+++ format/src/test/java/mathtools/format/RoundingTest.java
@@ -1,7 +1,7 @@
-package mathtools.numbers.format;
+package mathtools.format;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static mathtools.numbers.format.Rounding.round;
+import static mathtools.format.Rounding.round;
 
 import org.junit.jupiter.api.Test;
